test: use ContextConfiguration instead of profiles for DB
