+++ dhis-2/dhis-support/dhis-support-external/src/main/java/org/hisp/dhis/external/conf/DefaultDhisConfigurationProvider.java
@@ -58,7 +58,6 @@ import org.hisp.dhis.external.location.LocationManager;
 import org.hisp.dhis.external.location.LocationManagerException;
 import org.hisp.dhis.external.util.LogOnceLogger;
 import org.slf4j.event.Level;
-import org.springframework.context.annotation.Profile;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.support.PropertiesLoaderUtils;
 import org.springframework.stereotype.Component;
@@ -66,10 +65,6 @@ import org.springframework.stereotype.Component;
 /**
  * @author Lars Helge Overland
  */
-@Profile({
-  "!test-h2",
-  "!test-postgres",
-})
 @Component("dhisConfigurationProvider")
 @Slf4j
 public class DefaultDhisConfigurationProvider extends LogOnceLogger
