+++ dhis-2/dhis-support/dhis-support-test/src/main/resources/log4j2-test.xml
@@ -47,5 +47,9 @@
         <!-- <Logger name="org.springframework.test.context.cache" level="debug">-->
         <!--     <AppenderRef ref="console"/>-->
         <!-- </Logger>-->
+        <!-- Uncomment to debug Spring bean creation/test configuration -->
+        <!--<Logger name="org.springframework.beans.factory.support.DefaultListableBeanFactory" level="debug" additivity="false">
+                <AppenderRef ref="console"/>
+        </Logger>-->
     </Loggers>
 </Configuration>
