+++ dhis-2/dhis-test-integration/pom.xml
@@ -326,7 +326,6 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-dependency-plugin</artifactId>
-        <!--                Warning:  Unused declared dependencies found:-->
         <configuration>
           <failOnWarning>true</failOnWarning>
           <ignoredNonTestScopedDependencies>
