+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/test/integration/PostgresIntegrationTestBase.java
@@ -29,7 +29,8 @@ package org.hisp.dhis.test.integration;
 
 import org.hisp.dhis.test.IntegrationTest;
 import org.hisp.dhis.test.IntegrationTestBase;
-import org.springframework.test.context.ActiveProfiles;
+import org.hisp.dhis.test.config.PostgresTestConfig;
+import org.springframework.test.context.ContextConfiguration;
 
 /**
  * Base class for all Spring based integration tests which use a Postgres DB running in a Docker
@@ -38,5 +39,5 @@ import org.springframework.test.context.ActiveProfiles;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 @IntegrationTest
-@ActiveProfiles(profiles = {"test-postgres"})
+@ContextConfiguration(classes = {PostgresTestConfig.class})
 public abstract class PostgresIntegrationTestBase extends IntegrationTestBase {}
