+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/test/webapi/AuthenticationApiTestBase.java
@@ -27,7 +27,7 @@
  */
 package org.hisp.dhis.test.webapi;
 
-import org.hisp.dhis.test.config.H2DhisTestConfig;
+import org.hisp.dhis.test.config.H2TestConfig;
 import org.hisp.dhis.test.webapi.AuthenticationApiTestBase.AuthConfigProviderConfig;
 import org.hisp.dhis.webapi.security.config.WebMvcConfig;
 import org.junit.jupiter.api.BeforeEach;
@@ -54,7 +54,7 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 @ContextConfiguration(
     inheritLocations = false,
     classes = {
-      H2DhisTestConfig.class,
+      H2TestConfig.class,
       WebMvcConfig.class,
       AuthConfigProviderConfig.class,
     })
