+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/security/config/PasswordEncoderConfig.java
@@ -29,7 +29,6 @@ package org.hisp.dhis.webapi.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
@@ -37,7 +36,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
  * @author Morten Svanæs <msvanaes@dhis2.org>
  */
 @Configuration
-@Order(903)
 public class PasswordEncoderConfig {
   @Bean
   public PasswordEncoder passwordEncoder() {
